Create QudevDisplacer meta-instrument. Relocate other class
